Update Google API v16

+++ bin/GoogleAdsCleanupServices.php
@@ -26,7 +26,7 @@ class GoogleAdsCleanupServices {
 	 *
 	 * @var string
 	 */
-	protected $version = 'V14';
+	protected $version = 'V16';
 
 	/**
 	 * @var Event Composer event.
@@ -49,10 +49,17 @@ class GoogleAdsCleanupServices {
 	protected $code_path = null;
 
 	/**
+	 * @var string[] List of Service to NOT remove even when usage is not found.
+	 */
+	protected $avoid_cleanup = [
+		'ExtensionFeedItem',
+	];
+
+	/**
 	 * Constructor.
 	 *
 	 * @param Event|null  $event Composer event.
-	 * @param string|null $path  Path of the Ads library.
+	 * @param string|null $path Path of the Ads library.
 	 */
 	public function __construct( Event $event = null, string $path = null ) {
 		$this->event     = $event;
@@ -92,7 +99,10 @@ class GoogleAdsCleanupServices {
 		$used = array_unique(
 			array_merge(
 				$this->find_used_pattern(
-					"use Google\\\\Ads\\\\GoogleAds\\\\{$this->version}\\\\Services\\\\([A-Za-z0-9]+)ServiceClient;"
+					"use Google\\\\Ads\\\\GoogleAds\\\\{$this->version}\\\\Services\\\\Client\\\\([A-Za-z0-9]+)ServiceClient;"
+				),
+				$this->find_used_pattern(
+					"use Google\\\\Ads\\\\GoogleAds\\\\{$this->version}\\\\Services\\\\([A-Za-z0-9]+);"
 				),
 				$this->find_used_pattern(
 					"use Google\\\\Ads\\\\GoogleAds\\\\{$this->version}\\\\Resources\\\\([A-Za-z0-9]+);"
@@ -111,6 +121,11 @@ class GoogleAdsCleanupServices {
 	 * @param string $service Service name.
 	 */
 	protected function remove_service( string $service ) {
+		if ( in_array( $service, $this->avoid_cleanup, true ) ) {
+			$this->output_text( "Skip Removing service {$service}" );
+			return;
+		}
+
 		$this->output_text( "Removing service {$service}" );
 		$this->src_path = '/src/Google/Ads/GoogleAds';
 
@@ -185,8 +200,9 @@ class GoogleAdsCleanupServices {
 	/**
 	 * Find a list of files in a path, including subdirectories, matching a pattern.
 	 *
-	 * @param string $path    Package path
+	 * @param string $path Package path
 	 * @param string $pattern Regex pattern to match
+	 *
 	 * @return array Matching files
 	 */
 	protected function get_dir_contents( $path, $pattern ) {
@@ -243,7 +259,7 @@ class GoogleAdsCleanupServices {
 	 * Find a specific filename pattern within the library.
 	 *
 	 * @param string      $pattern Regexp pattern to match.
-	 * @param string|null $suffix  Suffix to remove from filename.
+	 * @param string|null $suffix Suffix to remove from filename.
 	 *
 	 * @return array List of matched names.
 	 */
@@ -257,6 +273,7 @@ class GoogleAdsCleanupServices {
 		return array_map(
 			function ( $file ) use ( $suffix ) {
 				$name = pathinfo( $file, PATHINFO_FILENAME );
+
 				return $suffix ? $this->remove_suffix( $suffix, $name ) : $name;
 			},
 			$output
@@ -267,7 +284,7 @@ class GoogleAdsCleanupServices {
 	 * Optionally remove a suffix from a string.
 	 *
 	 * @param string $suffix Suffix to remove.
-	 * @param string $text   Text to remove the suffix from.
+	 * @param string $text Text to remove the suffix from.
 	 *
 	 * @return string
 	 */
@@ -282,7 +299,7 @@ class GoogleAdsCleanupServices {
 	 * @param string $name
 	 */
 	protected function remove_use_statement( string $file, string $name ) {
-		$pattern = '/\s*use\s+[\w\d\\\]+' . preg_quote( $name, '/' ) . ';/';
+		$pattern = '/\s*use\s+[\w\d\\\]+\\\{1}' . preg_quote( $name, '/' ) . ';/';
 		$this->remove_pattern( $file, $pattern );
 	}
 
@@ -296,6 +313,7 @@ class GoogleAdsCleanupServices {
 		$file = $this->file_path( $file );
 		if ( ! file_exists( $file ) ) {
 			$this->warning( sprintf( 'File does not exist: %s', $file ) );
+
 			return;
 		}
 
@@ -304,6 +322,7 @@ class GoogleAdsCleanupServices {
 		$pattern = '/function ' . preg_quote( $name, '/' ) . '\(/';
 		if ( ! preg_match( $pattern, $contents, $matches, PREG_OFFSET_CAPTURE ) ) {
 			$this->warning( sprintf( 'Function %s not found in %s', $name, $file ) );
+
 			return;
 		}
 
@@ -329,6 +348,7 @@ class GoogleAdsCleanupServices {
 
 		if ( false === $start || 0 > $start || $end >= $length ) {
 			$this->warning( sprintf( 'Function %s not found in %s', $name, $file ) );
+
 			return;
 		}
 
@@ -342,6 +362,7 @@ class GoogleAdsCleanupServices {
 
 		if ( empty( $new ) ) {
 			$this->warning( sprintf( 'Replace failed for function %s in %s', $name, $file ) );
+
 			return;
 		}
 
@@ -358,18 +379,21 @@ class GoogleAdsCleanupServices {
 		$file = $this->file_path( $file );
 		if ( ! file_exists( $file ) ) {
 			$this->warning( sprintf( 'File does not exist: %s', $file ) );
+
 			return;
 		}
 
 		$contents = file_get_contents( $file ); // phpcs:ignore WordPress.WP.AlternativeFunctions
 		if ( ! preg_match( $pattern, $contents, $matches ) ) {
 			$this->warning( sprintf( 'Pattern %s not found in %s', $pattern, $file ) );
+
 			return;
 		}
 
 		$new = preg_replace( $pattern, '', $contents );
 		if ( empty( $new ) ) {
 			$this->warning( sprintf( 'Replace failed for pattern %s in %s', $pattern, $file ) );
+
 			return;
 		}
 
@@ -385,6 +409,7 @@ class GoogleAdsCleanupServices {
 		$file = $this->file_path( $file );
 		if ( ! file_exists( $file ) ) {
 			$this->warning( sprintf( 'File does not exist: %s', $file ) );
+
 			return;
 		}
 
@@ -400,10 +425,26 @@ class GoogleAdsCleanupServices {
 		$directory = $this->file_path( $directory );
 		if ( ! is_dir( $directory ) ) {
 			$this->warning( sprintf( 'Directory does not exist: %s', $directory ) );
+
 			return;
 		}
 
-		rmdir( $directory ); // phpcs:ignore WordPress.WP.AlternativeFunctions.file_system_operations_rmdir
+		if ( $this->is_empty_directory( $directory ) ) {
+			rmdir( $directory ); // phpcs:ignore WordPress.WP.AlternativeFunctions.file_system_operations_rmdir
+		} else {
+			$this->warning( sprintf( 'Directory not empty: %s', $directory ) );
+		}
+	}
+
+	/**
+	 * Check if a directory is empty
+	 *
+	 * @param string $directory
+	 *
+	 * @return bool True if empty
+	 */
+	protected function is_empty_directory( string $directory ): bool {
+		return ( count( scandir( $directory ) ) === 2 );
 	}
 
 	/**
